Revamp input prefilling
- we can now prefill based on service name
- filter only TCP/UDP ports
- utilize the full raw API types in a custom function

This moved the prefill logic out of generic-attack-form into workspace-attacks.

The preview in workspace-attacks now also shows the actual data that is going to be sent via the API now.

The new architecture is a bit easier to understand than the old one I think.

Weird selection combinations now filter to the valid subset of values. Since we got a preview of full data for the user now, this is better than having them error when trying to submit.

Fixes disabled attacks based on input types, removes the need to hardcode the allowed attacks for which selection types.