Swapped the order in which ips and ports are iterated
Previously they were: `for ip in ips { for port in ports { ... } }`
This meant that an ip's high ports were scanned before the next ip's
low ports were which is unintuitive.
So the new order is: `for port in ports { for ip in ips {...} }`
This also balances the load over all hosts since the loop body is run
concurrent.